requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-gis>=1.1,<2.0
djangorestframework-simplejwt>=5.3,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
redis>=5.0,<7.0
gunicorn>=23.0,<24.0
